oops
(cherry picked from commit bf0227cf02d145da2020a08e072d850c6fd5fbc5)

+++ scripts/build-android-armv7a.sh
@@ -8,4 +8,8 @@ tar -xf clang+llvm-11.1.0-x86_64-linux-gnu-ubuntu-16.04.tar.xz
 export PATH="$PWD/clang+llvm-11.1.0-x86_64-linux-gnu-ubuntu-16.04/bin:$PATH"
 ./waf configure -T release --build-game=cstrike --prefix=Srceng-mod-launcher/android --android=armeabi-v7a-hard,host,21 --target=../armeabi-v7a --enable-opus --togles --disable-warns &&
 ./waf install --target=client,server,engine,vgui2,vguimatsurface --strip
-rm Srceng-mod-launcher/android/lib/armeabi-v7a/README.md
+
+if [ -e "Srceng-mod-launcher/android/lib/armeabi-v7a/README.md" ]; then
+	rm Srceng-mod-launcher/android/lib/armeabi-v7a/README.md
+fi
+
